requirements.txt
requests
packaging
pprint
PyYaml